requirements.txt
asgiref==3.4.1
astroid==2.8.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.9.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==13.14.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
ipython==7.30.0
isort==5.9.3
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
openapi-codec==1.3.2
packaging==21.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.23
psycopg2==2.9.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.12.1
